modify error and exception handling

+++ src/main/java/com/mbpt/peertutoringplatform/service/impl/ClassRoomServiceImpl.java
@@ -3,7 +3,7 @@ package com.mbpt.peertutoringplatform.service.impl;
 import com.mbpt.peertutoringplatform.dto.ClassRoomDTO;
 import com.mbpt.peertutoringplatform.dto.MentorDTO;
 import com.mbpt.peertutoringplatform.entity.ClassRoomEntity;
-import com.mbpt.peertutoringplatform.exception.ClassRoomException;
+import com.mbpt.peertutoringplatform.exception.ResourceNotFoundException;
 import com.mbpt.peertutoringplatform.mapper.ClassRoomEntityDTOMapper;
 import com.mbpt.peertutoringplatform.mapper.MentorEntityDTOMapper;
 import com.mbpt.peertutoringplatform.repository.ClassRoomRepository;
@@ -38,7 +38,7 @@ public class ClassRoomServiceImpl implements ClassRoomService {
     public ClassRoomDTO createClassRoom(String title, MultipartFile classImage) {
         log.info("Creating new classroom...");
 
-        if (classImage.isEmpty() || title.isBlank() ) {
+        if (classImage == null || classImage.isEmpty() || title == null || title.isBlank() ) {
             log.error("Failed to create classroom: input data is invalid.");
             throw new IllegalArgumentException("Classroom data are required");
         }
@@ -86,6 +86,9 @@ public class ClassRoomServiceImpl implements ClassRoomService {
     @Transactional(readOnly = true)
     public ClassRoomDTO findClassRoomById(Integer id) {
         log.info("Fetching classroom by ID: {} ...", id);
+
+        if(id == null) throw new IllegalArgumentException("Classroom id is required");
+
         return classRoomRepository.findById(id)
                 .map(classRoom -> {
                     log.info("Classroom found: {}", classRoom);
@@ -96,55 +99,37 @@ public class ClassRoomServiceImpl implements ClassRoomService {
                 })
                 .orElseThrow(() -> {
                     log.error("Classroom not found with ID: {} from data-source:{}", id, this.datasource);
-                    return new ClassRoomException("Classroom not found with ID: " + id);
+                    return new ResourceNotFoundException("Classroom not found with ID: " + id);
                 });
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<ClassRoomDTO> findClassRoomsByFilters(String title, String mentorName, Integer minCount, Integer maxCount) {
-        log.info("Searching classroom(s) by filtering...");
-
-        List<ClassRoomEntity> classRoomEntities = classRoomRepository.findAllClassRooms(title, mentorName, minCount, maxCount);
-
-        List<ClassRoomDTO> classRoomDTOList = classRoomEntities.stream()
-                .map(classRoomEntity -> {
-                    ClassRoomDTO classRoomDTO = ClassRoomEntityDTOMapper.map(classRoomEntity);
-                    classRoomDTO.setMentorDTO(MentorEntityDTOMapper.map(classRoomEntity.getMentorEntity()));
-                    return classRoomDTO;
-                }).toList();
-
-        log.info("Found {} classroom(s) from data-source: {}", classRoomDTOList.size(), this.datasource);
-        return classRoomDTOList;
     }
 
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public ClassRoomDTO updateClassRoomById(ClassRoomDTO classRoomDTO) {
-        log.info("Updating classroom...");
+    public ClassRoomDTO updateClassRoomById(Integer id, String title, MultipartFile classImage, String imageUrl) {
+        log.info("Updating classroom with id: {} ...", id);
 
-        if (classRoomDTO == null || classRoomDTO.getClassRoomId() == null) {
-            log.error("Failed to update classroom: DTO or classroom's ID is null.");
-            throw new IllegalArgumentException("Classroom ID must not be null for update.");
-        }
+        if(id == null || title == null) throw new IllegalArgumentException("Classroom id and title are required");
 
-        Integer classRoomId = classRoomDTO.getClassRoomId();
-
-        log.debug("Updating classroom with ID: {}", classRoomId);
-
-        ClassRoomEntity classRoomEntity = classRoomRepository.findById(classRoomId)
+        ClassRoomEntity classRoomEntity = classRoomRepository.findById(id)
                 .orElseThrow(() -> {
-                    log.error("Failed to update classroom: Classroom not found with ID: {}", classRoomId);
-                    return new ClassRoomException("Classroom not found with ID: " + classRoomId);
+                    log.error("Failed to update classroom: Classroom not found with ID: {}", id);
+                    return new ResourceNotFoundException("Classroom not found with ID: " + id);
                 });
 
-        classRoomEntity.setTitle(classRoomDTO.getTitle());
-        classRoomEntity.setEnrolledStudentCount(classRoomDTO.getEnrolledStudentCount());
-        classRoomEntity.setClassImage(classRoomDTO.getClassImage());
+        classRoomEntity.setTitle(title);
+
+        if (classImage != null && !classImage.isEmpty()) {
+            String newImageUrl = fileService.uploadImage(classImage);
+            classRoomEntity.setClassImage(newImageUrl);
+        } else if (imageUrl != null && !imageUrl.isBlank()) {
+            classRoomEntity.setClassImage(imageUrl);
+        }
 
         ClassRoomEntity updatedClassroomEntity = classRoomRepository.save(classRoomEntity);
-        log.info("Updated classroom with ID: {}", classRoomId);
+
+        log.info("Updated classroom with ID: {}", id);
+
         return ClassRoomEntityDTOMapper.map(updatedClassroomEntity);
     }
 
@@ -153,10 +138,13 @@ public class ClassRoomServiceImpl implements ClassRoomService {
     @Transactional(rollbackFor = Exception.class)
     public ClassRoomDTO deleteClassRoomById(Integer id) {
         log.info("Deleting classroom with ID: {} ...", id);
+
+        if(id == null) throw new IllegalArgumentException("Classroom id is required");
+
         ClassRoomEntity classRoomEntity = classRoomRepository.findById(id)
                 .orElseThrow(() -> {
                     log.error("Failed to delete classroom. Classroom not found with ID: {}", id);
-                    return new ClassRoomException("Failed to delete classroom. Classroom not found with ID: " + id);
+                    return new ResourceNotFoundException("Failed to delete classroom. Classroom not found with ID: " + id);
                 });
         classRoomRepository.deleteById(id);
         log.info("Deleted classroom with ID: {} ", id);
